requirements.txt
tensorflow
numpy
pandas
Pillow
matplotlib
scikit-learn   

streamlit




